requirements.txt
fastapi
uvicorn
pyjwt
python-dotenv
cryptography
pymysql
sqlalchemy
aiomysql
aiosqlite
pytest
pytest-asyncio
pydantic[email]
bcrypt